+++ lib/module/pick.ts
@@ -2,7 +2,6 @@
  * 获取对应索引的value
  * @param obj 对象
  * @param index 索引
- * @param parentKey 父级key
  * @returns
  * @see {@link https://yourhhh.github.io/zztoolDocument/#pick} API 文档
  * @example
@@ -12,30 +11,33 @@
 export function pick(
   obj: any,
   index: string,
-  parentKey: string = ""
+  // parentKey: string = "",
 ): Array<{ key: string; value: any }> {
   if (typeof obj !== "object" || obj === null) throw new Error("obj must be an object and not null");
   if (typeof index !== "string") throw new Error("index must be string");
-  if (typeof parentKey !== "string") throw new Error("parentKey must be string");
+  // if (typeof parentKey !== "string") throw new Error("parentKey must be string");
   const arr: { key: string; value: any }[] = [];
-  for (const key in obj) {
-    if (obj.hasOwnProperty(key)) {
-      const fullKey = parentKey ? `${parentKey}.${key}` : key;
-      const value = obj[key];
-
-      // 递归处理非 null 的对象类型
-      if (typeof value === "object" && value !== null) {
-        // 添加 null 检查
-        arr.push(...pick(value, index, fullKey));
-      }
-
-      // 收集匹配键的值（严格相等检查）
-      if (key === index) {
-        // 改为独立 if 语句，允许同时处理值和子对象
-        arr.push({
-          key: fullKey,
-          value,
-        });
+  const visited = new WeakSet<Object>();
+  // 暂存栈
+  const stack: Array<{ value: any; path: string }> = [{value:obj,path:""}];
+  while (stack.length > 0){
+    const {value:currentData,path} = stack.pop()!;
+    if (typeof currentData === "object" && currentData !== null) {
+      if (visited.has(currentData)) continue; // 避免循环引用
+      visited.add(currentData);
+      for (const key in currentData) {
+        if (currentData.hasOwnProperty(key)) {
+          const fullKey =  path ? `${path}.${key}` : key;
+          const value = currentData[key];
+          
+          if (typeof value === "object" && value !== null) {
+            // 入栈
+            stack.push({value,path:fullKey});
+          }
+          if (key === index) {
+            arr.push({key: fullKey, value});
+          }
+        }
       }
     }
   }