+++ lib/utils/public.ts
@@ -13,4 +13,5 @@ export function defaultCheckEmpty(value:any):boolean{
   if (typeof value === "number" || typeof value === "boolean") return false;
   if (Array.isArray(value)) return value.length === 0;
   if (typeof value === "object") return Object.keys(value).length === 0;
+  return false;
 }