+++ src/App.vue
@@ -1,38 +1,15 @@
-<script setup lang="ts">
-import { isEmpty,pick } from '../lib';
-import {generateLargeComplexObject} from "../script/generateData"
-function createBigData(size: number) {
-  const arr = [];
-  for (let i = 0; i < size; i++) {
-    arr.push({ id: i, name: `name_${i}`, info: { value: i % 2 === 0 ? "" : i } });
-  }
-  return arr;
-}
-const bigData = createBigData(10_000_000);
-
-console.time("isEmpty - 1000w");
-isEmpty(bigData, { returnKeys: false });
-console.timeEnd("isEmpty - 1000w");
+<script setup lang="js">
+import { pick } from '../lib';
+import { generateLargeComplexObject } from '../script/generateData';
 
 
-function chunkProcess<T>(data: T[], chunkSize: number, callback: (chunk: T[]) => void, done: () => void) {
-  let index = 0;
-  function next() {
-    const slice = data.slice(index, index + chunkSize);
-    if (slice.length > 0) {
-      callback(slice);
-      index += chunkSize;
-      setTimeout(next); // 把任务分片，避免长时间阻塞
-    } else {
-      done();
-    }
-  }
-  next();
-}
-
-// 用法
-// const bigData2 = createBigData(100_000_000);
-// console.time("chunk isEmpty");
+// // // 用法
+// const bigData2 = generateLargeComplexObject(1_000_000);
+const bigData2 = generateLargeComplexObject(100);
+console.log(bigData2)
+console.time("pick");
+console.log(pick(bigData2, "ags",'test'));
+console.timeEnd("pick");
 
 // chunkProcess(bigData2, 100000, (chunk) => {
 //   isEmpty(chunk, { returnKeys: false });